+++ evmd/app/interfaces.go
@@ -4,8 +4,16 @@
 package app
 
 import (
+	erc20keeper "github.com/cosmos/evm/x/erc20/keeper"
 	feemarketkeeper "github.com/cosmos/evm/x/feemarket/keeper"
+	ibccallbackskeeper "github.com/cosmos/evm/x/ibc/callbacks/keeper"
+	transferkeeper "github.com/cosmos/evm/x/ibc/transfer/keeper"
+	precisebankkeeper "github.com/cosmos/evm/x/precisebank/keeper"
 	evmkeeper "github.com/cosmos/evm/x/vm/keeper"
+	ibckeeper "github.com/cosmos/ibc-go/v10/modules/core/keeper"
+
+	evidencekeeper "cosmossdk.io/x/evidence/keeper"
+	feegrantkeeper "cosmossdk.io/x/feegrant/keeper"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/types"
@@ -18,10 +26,11 @@ import (
 	mintkeeper "github.com/cosmos/cosmos-sdk/x/mint/keeper"
 	slashingkeeper "github.com/cosmos/cosmos-sdk/x/slashing/keeper"
 	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
-	"github.com/cosmos/ibc-go/v10/modules/core/keeper"
 )
 
-func (app App) GetIBCKeeper() *keeper.Keeper {
+// Getters of necessary keepers
+
+func (app App) GetIBCKeeper() *ibckeeper.Keeper {
 	return app.IBCKeeper
 }
 
@@ -76,4 +85,42 @@ func (app App) GetDistrKeeper() distrkeeper.Keeper {
 
 func (app App) GetConsensusParamsKeeper() consensuskeeper.Keeper {
 	return app.ConsensusParamsKeeper
+}
+
+// Getters of optional keepers
+//
+// These getters expose keepers that the minimum EVM app does not include,
+// and some of the existing tests may require one or more of these getters.
+// In such cases, you need to add the required keeper(s) to the app first,
+// and then implement the corresponding getter(s).
+func (app App) GetErc20Keeper() *erc20keeper.Keeper {
+	panic("implement me")
+}
+
+func (app App) SetErc20Keeper(keeper erc20keeper.Keeper) {
+	panic("implement me")
+}
+
+func (app App) GetEvidenceKeeper() *evidencekeeper.Keeper {
+	panic("implement me")
+}
+
+func (app App) GetPreciseBankKeeper() *precisebankkeeper.Keeper {
+	panic("implement me")
+}
+
+func (app App) GetFeeGrantKeeper() feegrantkeeper.Keeper {
+	panic("implement me")
+}
+
+func (app App) GetCallbackKeeper() ibccallbackskeeper.ContractKeeper {
+	panic("implement me")
+}
+
+func (app App) GetTransferKeeper() transferkeeper.Keeper {
+	panic("implement me")
+}
+
+func (app App) SetTransferKeeper(transferKeeper transferkeeper.Keeper) {
+	panic("implement me")
 }
